refactor(wave7c-step2): mechanically split judge and json_strict internals
