+++ crates/assay-evidence/src/json_strict/json_strict_internal/limits.rs
@@ -1,1 +1,3 @@
-//! Step2 scaffold: strict JSON limits boundary target.
+pub(crate) use crate::json_strict::errors::{
+    MAX_KEYS_PER_OBJECT, MAX_NESTING_DEPTH, MAX_STRING_LENGTH,
+};
