+++ crates/assay-evidence/src/json_strict/json_strict_internal/run.rs
@@ -1,1 +1,12 @@
-//! Step2 scaffold: optional json_strict orchestration entrypoint target.
+use crate::json_strict::errors::StrictJsonError;
+use serde::de::DeserializeOwned;
+
+pub(crate) fn from_str_strict_impl<T: DeserializeOwned>(s: &str) -> Result<T, StrictJsonError> {
+    validate_json_strict_impl(s)?;
+    Ok(serde_json::from_str(s)?)
+}
+
+pub(crate) fn validate_json_strict_impl(s: &str) -> Result<(), StrictJsonError> {
+    let mut validator = super::validate::JsonValidator::new(s);
+    validator.validate()
+}
