+++ crates/assay-evidence/src/json_strict/mod.rs
@@ -60,20 +60,15 @@ mod errors;
 mod json_strict_internal;
 mod scan;
 
-use serde::de::DeserializeOwned;
-use std::str::CharIndices;
-
-use errors::StrictJsonError as Err;
 pub use errors::{StrictJsonError, MAX_KEYS_PER_OBJECT, MAX_NESTING_DEPTH, MAX_STRING_LENGTH};
-use scan::parse_json_string;
+use serde::de::DeserializeOwned;
 
 /// Parse JSON with strict duplicate key rejection.
 ///
 /// Scans the JSON for duplicate keys at any nesting level before deserializing.
 /// This ensures semantic consistency across different JSON parsers.
 pub fn from_str_strict<T: DeserializeOwned>(s: &str) -> Result<T, StrictJsonError> {
-    validate_json_strict(s)?;
-    Ok(serde_json::from_str(s)?)
+    json_strict_internal::run::from_str_strict_impl(s)
 }
 
 /// Validate JSON string for security issues without deserializing.
@@ -82,270 +77,8 @@ pub fn from_str_strict<T: DeserializeOwned>(s: &str) -> Result<T, StrictJsonErro
 /// - Duplicate keys at any nesting level
 /// - Lone surrogates in unicode escapes
 pub fn validate_json_strict(s: &str) -> Result<(), StrictJsonError> {
-    let mut validator = JsonValidator::new(s);
-    validator.validate()
+    json_strict_internal::run::validate_json_strict_impl(s)
 }
-
-/// Internal validator state machine.
-struct JsonValidator<'a> {
-    chars: std::iter::Peekable<CharIndices<'a>>,
-    key_tracker: dupkeys::ObjectKeyTracker,
-    current_path: String,
-    depth: usize,
-    _phantom: std::marker::PhantomData<&'a str>,
-}
-
-impl<'a> JsonValidator<'a> {
-    fn new(input: &'a str) -> Self {
-        Self {
-            chars: input.char_indices().peekable(),
-            key_tracker: dupkeys::ObjectKeyTracker::new(),
-            current_path: String::new(),
-            depth: 0,
-            _phantom: std::marker::PhantomData,
-        }
-    }
-
-    fn validate(&mut self) -> Result<(), StrictJsonError> {
-        self.skip_whitespace();
-        self.validate_value()?;
-        Ok(())
-    }
-
-    fn validate_value(&mut self) -> Result<(), StrictJsonError> {
-        self.skip_whitespace();
-
-        match self.peek_char() {
-            Some('{') => self.validate_object(),
-            Some('[') => self.validate_array(),
-            Some('"') => self.validate_string().map(|_| ()),
-            Some(c) if c == '-' || c.is_ascii_digit() => self.validate_number(),
-            Some('t') | Some('f') => self.validate_bool(),
-            Some('n') => self.validate_null(),
-            Some(c) => Err(Err::ParseError(
-                serde_json::from_str::<()>(&format!("unexpected char '{}'", c)).unwrap_err(),
-            )),
-            None => Ok(()),
-        }
-    }
-
-    fn validate_object(&mut self) -> Result<(), StrictJsonError> {
-        self.expect_char('{')?;
-        self.skip_whitespace();
-
-        self.depth += 1;
-        if self.depth > MAX_NESTING_DEPTH {
-            return Err(Err::NestingTooDeep { depth: self.depth });
-        }
-
-        let path = self.current_path.clone();
-        self.key_tracker.enter_object(path);
-
-        if self.peek_char() == Some('}') {
-            self.next_char();
-            self.key_tracker.exit_object();
-            self.depth -= 1;
-            return Ok(());
-        }
-
-        loop {
-            self.skip_whitespace();
-
-            let key = self.validate_string()?;
-            self.key_tracker.push_key(key.clone())?;
-
-            self.skip_whitespace();
-            self.expect_char(':')?;
-
-            let old_path = self.current_path.clone();
-            self.current_path = if self.current_path.is_empty() {
-                format!("/{}", key)
-            } else {
-                format!("{}/{}", self.current_path, key)
-            };
-
-            self.validate_value()?;
-
-            self.current_path = old_path;
-
-            self.skip_whitespace();
-
-            match self.peek_char() {
-                Some(',') => {
-                    self.next_char();
-                }
-                Some('}') => {
-                    self.next_char();
-                    self.key_tracker.exit_object();
-                    self.depth -= 1;
-                    return Ok(());
-                }
-                _ => {
-                    return Err(Err::ParseError(
-                        serde_json::from_str::<()>("expected ',' or '}'").unwrap_err(),
-                    ))
-                }
-            }
-        }
-    }
-
-    fn validate_array(&mut self) -> Result<(), StrictJsonError> {
-        self.expect_char('[')?;
-        self.skip_whitespace();
-
-        self.depth += 1;
-        if self.depth > MAX_NESTING_DEPTH {
-            return Err(Err::NestingTooDeep { depth: self.depth });
-        }
-
-        if self.peek_char() == Some(']') {
-            self.next_char();
-            self.depth -= 1;
-            return Ok(());
-        }
-
-        let mut index = 0;
-        loop {
-            let old_path = self.current_path.clone();
-            self.current_path = format!("{}/{}", self.current_path, index);
-
-            self.validate_value()?;
-
-            self.current_path = old_path;
-            index += 1;
-
-            self.skip_whitespace();
-
-            match self.peek_char() {
-                Some(',') => {
-                    self.next_char();
-                    self.skip_whitespace();
-                }
-                Some(']') => {
-                    self.next_char();
-                    self.depth -= 1;
-                    return Ok(());
-                }
-                _ => {
-                    return Err(Err::ParseError(
-                        serde_json::from_str::<()>("expected ',' or ']'").unwrap_err(),
-                    ))
-                }
-            }
-        }
-    }
-
-    fn validate_string(&mut self) -> Result<String, StrictJsonError> {
-        parse_json_string(&mut self.chars)
-    }
-
-    fn validate_number(&mut self) -> Result<(), StrictJsonError> {
-        if self.peek_char() == Some('-') {
-            self.next_char();
-        }
-
-        match self.peek_char() {
-            Some('0') => {
-                self.next_char();
-            }
-            Some(c) if c.is_ascii_digit() => {
-                while self.peek_char().is_some_and(|c| c.is_ascii_digit()) {
-                    self.next_char();
-                }
-            }
-            _ => {
-                return Err(Err::ParseError(
-                    serde_json::from_str::<()>("expected digit").unwrap_err(),
-                ))
-            }
-        }
-
-        if self.peek_char() == Some('.') {
-            self.next_char();
-            if !self.peek_char().is_some_and(|c| c.is_ascii_digit()) {
-                return Err(Err::ParseError(
-                    serde_json::from_str::<()>("expected digit after decimal").unwrap_err(),
-                ));
-            }
-            while self.peek_char().is_some_and(|c| c.is_ascii_digit()) {
-                self.next_char();
-            }
-        }
-
-        if self.peek_char().is_some_and(|c| c == 'e' || c == 'E') {
-            self.next_char();
-            if self.peek_char().is_some_and(|c| c == '+' || c == '-') {
-                self.next_char();
-            }
-            if !self.peek_char().is_some_and(|c| c.is_ascii_digit()) {
-                return Err(Err::ParseError(
-                    serde_json::from_str::<()>("expected digit in exponent").unwrap_err(),
-                ));
-            }
-            while self.peek_char().is_some_and(|c| c.is_ascii_digit()) {
-                self.next_char();
-            }
-        }
-
-        Ok(())
-    }
-
-    fn validate_bool(&mut self) -> Result<(), StrictJsonError> {
-        if self.consume_keyword("true") || self.consume_keyword("false") {
-            Ok(())
-        } else {
-            Err(Err::ParseError(
-                serde_json::from_str::<()>("expected bool").unwrap_err(),
-            ))
-        }
-    }
-
-    fn validate_null(&mut self) -> Result<(), StrictJsonError> {
-        if self.consume_keyword("null") {
-            Ok(())
-        } else {
-            Err(Err::ParseError(
-                serde_json::from_str::<()>("expected null").unwrap_err(),
-            ))
-        }
-    }
-
-    fn consume_keyword(&mut self, keyword: &str) -> bool {
-        let remaining: String = self.chars.clone().map(|(_, c)| c).collect();
-        if remaining.starts_with(keyword) {
-            for _ in keyword.chars() {
-                self.next_char();
-            }
-            true
-        } else {
-            false
-        }
-    }
-
-    fn skip_whitespace(&mut self) {
-        while self.peek_char().is_some_and(|c| c.is_ascii_whitespace()) {
-            self.next_char();
-        }
-    }
-
-    fn peek_char(&mut self) -> Option<char> {
-        self.chars.peek().map(|(_, c)| *c)
-    }
-
-    fn next_char(&mut self) -> Option<(usize, char)> {
-        self.chars.next()
-    }
-
-    fn expect_char(&mut self, expected: char) -> Result<(), StrictJsonError> {
-        match self.next_char() {
-            Some((_, c)) if c == expected => Ok(()),
-            _ => Err(Err::ParseError(
-                serde_json::from_str::<()>(&format!("expected '{}'", expected)).unwrap_err(),
-            )),
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
